closeappbackend/requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.8
Django==3.2.9
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.3
lxml==4.6.4
pytz==2021.3
requests==2.26.0
soupsieve==2.3.1
sqlparse==0.4.2
urllib3==1.26.7
whitenoise==5.3.0
